Implemented a Polyfit algorithm using numpy
Implemented a polyfit algorithm to figure out best m value to fit the data. I then calculated all the MSE values of each of the m's and stored them in a dictionary. I dynamically chose the right m based on the lowest m value. Then I tried to vary the noise and number of samples to see how that would affect the overall structure of the data and the MSE values and predictions.